+++ src/main/java/net/geforcemods/securitycraft/mixin/camera/ServerPlayerMixin.java
@@ -1,20 +1,38 @@
 package net.geforcemods.securitycraft.mixin.camera;
 
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import com.llamalad7.mixinextras.injector.v2.WrapWithCondition;
 
+import net.geforcemods.securitycraft.entity.camera.SecurityCamera;
 import net.geforcemods.securitycraft.util.PlayerUtils;
 import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.entity.Entity;
 
-/**
- * Makes sure the server does not move the player mounting a camera to the camera's position
- */
 @Mixin(value = ServerPlayer.class, priority = 1100)
-public class ServerPlayerMixin {
+public abstract class ServerPlayerMixin {
+	@Shadow
+	public abstract Entity getCamera();
+
+	/**
+	 * Makes sure the server does not move the player mounting a camera to the camera's position
+	 */
 	@WrapWithCondition(method = "tick", at = @At(value = "INVOKE", target = "Lnet/minecraft/server/level/ServerPlayer;absSnapTo(DDDFF)V"))
 	private boolean securitycraft$shouldMoveTo(ServerPlayer player, double x, double y, double z, float yaw, float pitch) {
 		return !PlayerUtils.isPlayerMountedOnCamera(player);
+	}
+
+	/**
+	 * Ensures that players in spectator mode are able to see players that are currently viewing a camera, because players
+	 * that are spectating another entity are usually invisible for other spectators
+	 */
+	@Inject(method = "broadcastToPlayer", at = @At("HEAD"), cancellable = true)
+	private void securitycraft$broadcastPlayerViewingCamera(ServerPlayer player, CallbackInfoReturnable<Boolean> cir) {
+		if (getCamera() instanceof SecurityCamera && player.isSpectator())
+			cir.setReturnValue(true);
 	}
 }